cf-remote: Split out ssh module
Avoids cyclic import

Changelog: None
